[Feat] 무한스크롤 로직 util 함수로 분리
- 무한스크롤 관련 로직 util 함수 생성하여 코드 분리
- 가독성 증진 및 유지보수 효율화 목적

Issues #19
